+++ client/lib/performance-tracking/performance-tracker-start.js
@@ -5,7 +5,7 @@ import { startPerformanceTracking } from './lib';
 
 export const performanceTrackerStart = ( pageName ) => {
 	return ( context, next ) => {
-		startPerformanceTracking( pageName, context.init );
+		startPerformanceTracking( pageName, context.init ?? false );
 		next();
 	};
 };
