+++ client/lib/performance-tracking/test/performance-tracker-start.js
@@ -1,41 +1,14 @@
-/**
- * External dependencies
- */
-import { start } from '@automattic/browser-data-collector';
+jest.mock( '../lib', () => ( {
+	startPerformanceTracking: jest.fn(),
+} ) );
 
 /**
  * Internal dependencies
  */
-import config from 'config';
 import { performanceTrackerStart } from '../performance-tracker-start';
-import { abtest } from 'lib/abtest';
-
-jest.mock( 'config', () => ( {
-	isEnabled: jest.fn(),
-} ) );
-jest.mock( '@automattic/browser-data-collector', () => ( {
-	start: jest.fn(),
-} ) );
-jest.mock( 'lib/abtest', () => ( {
-	abtest: jest.fn(),
-} ) );
-
-const withFeatureEnabled = () =>
-	config.isEnabled.mockImplementation( ( key ) => key === 'rum-tracking/logstash' );
-const withFeatureDisabled = () =>
-	config.isEnabled.mockImplementation( ( key ) => key !== 'rum-tracking/logstash' );
-const withABTestEnabled = () =>
-	abtest.mockImplementation( ( test ) =>
-		test === 'rumDataCollection' ? 'collectData' : 'noData'
-	);
-const withABTestDisabled = () => abtest.mockImplementation( () => 'noData' );
+import { startPerformanceTracking } from '../lib';
 
 describe( 'performance-tracker-start', () => {
-	beforeEach( () => {
-		withFeatureEnabled();
-		withABTestEnabled();
-	} );
-
 	afterEach( () => {
 		jest.resetAllMocks();
 	} );
@@ -49,55 +22,21 @@ describe( 'performance-tracker-start', () => {
 		expect( next ).toHaveBeenCalled();
 	} );
 
-	it( 'starts measuring when the feature is enabled', () => {
-		performanceTrackerStart( 'pageName' )( {}, jest.fn() );
-
-		expect( start ).toHaveBeenCalled();
-	} );
-
-	it( 'do not start measuring when the config flag is off', () => {
-		withFeatureDisabled();
-
-		performanceTrackerStart( 'pageName' )( {}, jest.fn() );
-
-		expect( start ).not.toHaveBeenCalled();
-	} );
-
-	it( 'do not start measuring when the abtest is disabled', () => {
-		withABTestDisabled();
-
-		performanceTrackerStart( 'pageName' )( {}, jest.fn() );
-
-		expect( start ).not.toHaveBeenCalled();
-	} );
-
 	it( 'uses the name of the page', () => {
 		performanceTrackerStart( 'pageName' )( {}, jest.fn() );
 
-		expect( start ).toHaveBeenCalledWith( 'pageName', expect.anything() );
-	} );
-
-	it( 'if the page name is not provided, uses "calypso" as the name', () => {
-		performanceTrackerStart()( {}, jest.fn() );
-
-		expect( start ).toHaveBeenCalledWith( 'calypso', expect.anything() );
+		expect( startPerformanceTracking ).toHaveBeenCalledWith( 'pageName', expect.anything() );
 	} );
 
 	it( 'measures fullPageLoad if the page is an initial navigation', () => {
 		performanceTrackerStart( 'pageName' )( { init: true }, jest.fn() );
 
-		expect( start ).toHaveBeenCalledWith( expect.anything(), { fullPageLoad: true } );
+		expect( startPerformanceTracking ).toHaveBeenCalledWith( expect.anything(), true );
 	} );
 
 	it( 'does not measure fullPageLoad if the page is not an initial navigation', () => {
 		performanceTrackerStart( 'pageName' )( { init: false }, jest.fn() );
 
-		expect( start ).toHaveBeenCalledWith( expect.anything(), { fullPageLoad: false } );
-	} );
-
-	it( 'does not measure fullPageLoad by default', () => {
-		performanceTrackerStart( 'pageName' )( {}, jest.fn() );
-
-		expect( start ).toHaveBeenCalledWith( expect.anything(), { fullPageLoad: false } );
+		expect( startPerformanceTracking ).toHaveBeenCalledWith( expect.anything(), false );
 	} );
 } );
