+++ client/lib/performance-tracking/test/use-performance-tracker-stop.js
@@ -3,15 +3,13 @@
  */
 import { useLayoutEffect } from 'react';
 import { useSelector } from 'react-redux';
-import { stop } from '@automattic/browser-data-collector';
 
 /**
  * Internal dependencies
  */
 import { getSectionName } from 'state/ui/selectors';
-import config from 'config';
 import { usePerformanceTrackerStop } from '../use-performance-tracker-stop';
-import { getABTestVariation } from 'lib/abtest';
+import { stopPerformanceTracking } from '../lib';
 
 jest.mock( 'react', () => ( {
 	useLayoutEffect: jest.fn(),
@@ -19,28 +17,12 @@ jest.mock( 'react', () => ( {
 jest.mock( 'react-redux', () => ( {
 	useSelector: jest.fn(),
 } ) );
-jest.mock( '@automattic/browser-data-collector', () => ( {
-	stop: jest.fn(),
+jest.mock( '../lib', () => ( {
+	stopPerformanceTracking: jest.fn(),
 } ) );
 jest.mock( 'state/ui/selectors', () => ( {
 	getSectionName: jest.fn(),
 } ) );
-jest.mock( 'config', () => ( {
-	isEnabled: jest.fn(),
-} ) );
-jest.mock( 'lib/abtest', () => ( {
-	getABTestVariation: jest.fn(),
-} ) );
-
-const withFeatureEnabled = () =>
-	config.isEnabled.mockImplementation( ( key ) => key === 'rum-tracking/logstash' );
-const withFeatureDisabled = () =>
-	config.isEnabled.mockImplementation( ( key ) => key !== 'rum-tracking/logstash' );
-const withABTestEnabled = () =>
-	getABTestVariation.mockImplementation( ( test ) =>
-		test === 'rumDataCollection' ? 'collectData' : 'noData'
-	);
-const withABTestDisabled = () => getABTestVariation.mockImplementation( () => 'noData' );
 
 describe( 'usePerfomranceTrackerStop hook', () => {
 	let originalRequestAnimationFrame;
@@ -50,8 +32,6 @@ describe( 'usePerfomranceTrackerStop hook', () => {
 		global.requestAnimationFrame = jest.fn( ( fn ) => fn() );
 		useLayoutEffect.mockImplementation( ( fn ) => fn() );
 		useSelector.mockImplementation( ( selector ) => selector() );
-		withFeatureEnabled();
-		withABTestEnabled();
 	} );
 
 	afterEach( () => {
@@ -59,28 +39,12 @@ describe( 'usePerfomranceTrackerStop hook', () => {
 		global.requestAnimationFrame = originalRequestAnimationFrame;
 	} );
 
-	it( 'does not stop measuring if the feature is off', () => {
-		withFeatureDisabled();
-
-		usePerformanceTrackerStop();
-
-		expect( stop ).not.toHaveBeenCalled();
-	} );
-
-	it( 'does not stop measuring if the abtest is disabled', () => {
-		withABTestDisabled();
-
-		usePerformanceTrackerStop();
-
-		expect( stop ).not.toHaveBeenCalled();
-	} );
-
 	it( 'gets the page name from the state', () => {
 		getSectionName.mockImplementation( () => 'pageName' );
 
 		usePerformanceTrackerStop();
 
-		expect( stop ).toHaveBeenCalledWith( 'pageName' );
+		expect( stopPerformanceTracking ).toHaveBeenCalledWith( 'pageName' );
 	} );
 
 	it( 'calls stop using useLayoutEffect and requestAnimationFrame', () => {
